+++ src/icv_sitemaps/__init__.py
@@ -2,7 +2,7 @@
 icv-sitemaps — Scalable sitemap generation and web discovery infrastructure for Django.
 """
 
-__version__ = "0.7.0"
+__version__ = "0.8.0"
 
 default_app_config = "icv_sitemaps.apps.IcvSitemapsConfig"